ADD test for Service Classes, minor fixes

+++ src/main/java/com/ramis/messenger/service/ChatService.java
@@ -1,10 +1,8 @@
 package com.ramis.messenger.service;
 
 import com.ramis.messenger.models.Chat;
-import com.ramis.messenger.models.Message;
 import com.ramis.messenger.models.User;
 import com.ramis.messenger.repository.ChatRepository;
-import com.ramis.messenger.repository.MessageRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,14 +11,16 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class ChatService {
     private final ChatRepository chatRepository;
-    private final MessageRepository messageRepository;
 
-
+    @Transactional
     public Chat createChat(User user, String chatName) {
+        if (user == null) {
+            throw new IllegalArgumentException("User can't be null");
+        }
         Chat chat = new Chat();
         chat.setName(chatName);
         chat.setCreator(user);
@@ -28,16 +28,14 @@ public class ChatService {
         return chatRepository.save(chat);
     }
 
-    //TODO: make
-    public List<Chat> getChats(User user) {
-        return null;
-    }
-
     /*
      *  Name can confuse, but it means that users for this chat is loaded.
      *  You can think that it's finding chats where List<User> not null, it's not.
      */
     public Chat getChatWithUsers(Long chatId) {
+        if (chatId == null || chatId < 1) {
+            throw new IllegalArgumentException("Chat id can't be null or less than 1");
+        }
         return chatRepository.getChatById(chatId).orElseThrow(
                 () -> new EntityNotFoundException("Chat with id " + chatId + " not found"));
     }
@@ -50,9 +48,10 @@ public class ChatService {
         return chats;
     }
 
+    @Transactional
     public void delete(Long id) {
-        if (id == null) {
-            throw new RuntimeException("ID cannot be null");
+        if (id == null || id < 1) {
+            throw new IllegalArgumentException("ID cannot be null or less than 1");
         }
         chatRepository.deleteById(id);
     }
